==> 02-data-description.Rmd <==
---
title: "03-data-descriptive"
author: "Sophie Hensgen"
date: "5/21/2021"
output: html_document
---

```{r setup, include=FALSE}
library(readr)
library(tidyverse)
library(dplyr)
library(ggplot2)
library(tidyr)
library(data.table)
library(sem)
library(reshape2)
library(xtable)
library(psych)
library(stargazer)
```

## Load Data Set

```{r Loading Data Set, include = FALSE}

df_stigma_wide <- read_csv("df_stigma_wide.csv")
df_stigma_long <- read_csv("df_stigma_long.csv")

```

## Descriptive

```{r descriptive social demo, include = FALSE}

### Summary with sds long and sds short

# soon subset of the data - only important variables should be included, such as age, sex etc. for short term
df_stigma_subset_long <- df_stigma_long %>%
  select(gen_sds_shortterm, gen_herkunft, gen_politik, gen_situation, gen_mc_largegroups, gen_mc_smallgroups, gen_mc_bars, gen_mc_none, gen_impfung, SD100, gen_alter, gen_bund, gen_schule, gen_schule_q, vignr, id)

# spread data set in long format, as we are only interested in the participant Level for the summary
df_stigma_subset_wide <- spread(df_stigma_subset_long, vignr, gen_sds_shortterm)

# soon subset of the data - only important variables should be included, such as age, sex etc. for long term
df_stigma_subset_long2 <- df_stigma_long %>%
  select(gen_sds_longterm, gen_herkunft, gen_politik, gen_situation, gen_mc_largegroups, gen_mc_smallgroups, gen_mc_bars, gen_mc_none, gen_impfung, SD100, gen_alter, gen_bund, gen_schule, gen_schule_q, vignr, id)

# spread data set in long format, as we are only interested in the participant Level for the summary
df_stigma_subset_wide2 <- spread(df_stigma_subset_long2, vignr, gen_sds_longterm)

# Combining both data sets 
df_stigma_subset_wide <- df_stigma_subset_wide %>%
  full_join(df_stigma_subset_wide2, by = c("id", "gen_herkunft", "gen_politik", "gen_situation", "gen_mc_largegroups", "gen_mc_smallgroups", "gen_mc_bars", "gen_mc_none", "gen_impfung", "SD100", "gen_alter", "gen_bund", "gen_schule", "gen_schule_q"))

# Rename new columns
colnames(df_stigma_subset_wide) <- c("gen_herkunft", "gen_politik", "gen_situation", "gen_mc_largegroups", "gen_mc_smallgroups", "gen_mc_bars", "gen_mc_none", "gen_impfung", "SD100", "gen_alter", "gen_bund", "gen_schule", "gen_schule_q", "id", "gen_sds_shortterm_v1", "gen_sds_shortterm_v2", "gen_sds_shortterm_v3", "gen_sds_shortterm_v4", "gen_sds_longterm_v1", "gen_sds_longterm_v2", "gen_sds_longterm_v3", "gen_sds_longterm_v4")

## What to include in Summary
# Question:
#   Which measurements should be used?
#   -   Min and Max
#   -   Mean
#   -   N
#   -   Standard Deviation
# Important!! Only for variables which are NOT Categorical! for those we look into the individual percentage

# Creating a latex Table
stargazer(data.frame(df_stigma_subset_wide), type = "latex", summary = TRUE,summary.stat = c("n", "min", "max", "mean", "median", "sd"),  out = "summary-1.tex", digits=2)

# Calculate the percentages for each factor in these Variables!
Gender <- (as.data.frame(table(df_stigma_subset_wide$SD100)) %>% rename(Count=1,Freq=2) %>% mutate(Perc=100*Freq/sum(Freq)))
Age <- (as.data.frame(table(df_stigma_subset_wide$gen_alter)) %>% rename(Count=1,Freq=2) %>% mutate(Perc=100*Freq/sum(Freq)))
State <- (as.data.frame(table(df_stigma_subset_wide$gen_bund)) %>% rename(Count=1,Freq=2) %>% mutate(Perc=100*Freq/sum(Freq)))
School <- (as.data.frame(table(df_stigma_subset_wide$gen_schule_q)) %>% rename(Count=1,Freq=2) %>% mutate(Perc=100*Freq/sum(Freq)))
Migration <- (as.data.frame(table(df_stigma_subset_wide$gen_herkunft)) %>% rename(Count=1,Freq=2) %>% mutate(Perc=100*Freq/sum(Freq)))
Politics <- (as.data.frame(table(df_stigma_subset_wide$gen_politik)) %>% rename(Count=1,Freq=2) %>% mutate(Perc=100*Freq/sum(Freq)))
Vaccine <- (as.data.frame(table(df_stigma_subset_wide$gen_impfung)) %>% rename(Count=1,Freq=2) %>% mutate(Perc=100*Freq/sum(Freq)))

# Combining all rows to a data set
df_cat <- bind_rows(Gender, Age, State, School, Migration, Politics, Vaccine)

stargazer(data.frame(df_cat), type = "latex", 
       summary = FALSE,
       title = "Quotation Table",
       label="tab4",
       covariate.labels = c("", "Label", "Frequency","Percentages"),
       digits = 1)
       #dep.var.labels   = c("Social Distance", "Short Term - Social Distance", "Long Term - Social Distance"), 
       #table.placement = "H",
       #column.labels = c("Count", "Frequency", "Percentatage"))
```

```{r normal distribution - dependent variable}

### Normal Distribution

hist(df_stigma_long$gen_sds)
hist(df_stigma_long$gen_sds_longterm)
hist(df_stigma_long$gen_sds_shortterm)

hist(df_stigma_long$SDS100)
hist(df_stigma_long$SDS200)
hist(df_stigma_long$SDS300)
hist(df_stigma_long$SDS400)
hist(df_stigma_long$SDS500)
hist(df_stigma_long$SDS600)


### Interpretation

# we see spikes at 1, 2, 3 and 4 as predicted
# The index variable including all 4 Variables has a similar look as a normal distribution, however, 1 is much higher than the other values. Which did not show, when Florian tried it. 
# This is similar to the other index variables and the individual variables

```

```{r Vignette Differences}

### Vignette Mean Differences

df.test <- na.omit(df_stigma_long)

df.test %>%
  group_by(vignr) %>%
  summarize(meansds = mean(gen_sds))

df.test %>%
  group_by(vignr) %>%
  summarize(meanssds_long = mean(gen_sds_shortterm))

df.test %>%
  group_by(vignr) %>%
  summarize(meanssds_long = mean(gen_sds_longterm))

### Interpretation

# the mean of the individual Vignettes are quite similar and therefore should not be a problem

```

###########################################################################################################################################
##                                                                                                                                       ##
##                                                 Factor Analysis                                                                       ##
##                                                                                                                                       ##
###########################################################################################################################################


## Factor Analysis Social Distance Scale

```{r Factor analysis - Social Distance Scale}

# Subset data set

stigma.sds <- df_stigma_long %>%
  select(SDS100, SDS200, SDS300, SDS400, SDS500, SDS600)

### Factor Analysis

stigma.sds <- na.omit(stigma.sds)
stigma.sds.fa <- factanal(stigma.sds, factors = 2)


### Interpretation the results

apply(stigma.sds.fa$loadings^2,1,sum) # commonality (should be high)
1 - apply(stigma.sds.fa$loadings^2,1,sum) # uniqueness (should be low)

# factors are worth keeping, as their SS loadings are greater than 1 (Kaiser's rule)


### Residual Matrix

Lambda <- stigma.sds.fa$loadings
Psi <- diag(stigma.sds.fa$uniquenesses)
S <- stigma.sds.fa$correlation
Sigma <- Lambda %*% t(Lambda) + Psi
round(S - Sigma, 6)

```

```{r Factor analysis - Social Distance Scale - graphs}

### Interpretation Factors

stigma.sds.fa.none <- factanal(stigma.sds, factors = 2, rotation = "none")
stigma.sds.fa.varimax <- factanal(stigma.sds, factors = 2, rotation = "varimax")
stigma.sds.fa.promax <- factanal(stigma.sds, factors = 2, rotation = "promax")

###########################################################################################################################################

par(mfrow = c(1,3))
plot(stigma.sds.fa.none$loadings[,1], 
     stigma.sds.fa.none$loadings[,2],
     xlab = "Factor 1", 
     ylab = "Factor 2", 
     ylim = c(-1,1),
     xlim = c(-1,1),
     main = "No rotation")
abline(h = 0, v = 0)

###########################################################################################################################################

plot(stigma.sds.fa.varimax$loadings[,1], 
     stigma.sds.fa.varimax$loadings[,2],
     xlab = "Factor 1", 
     ylab = "Factor 2", 
     ylim = c(-1,1),
     xlim = c(-1,1),
     main = "Varimax rotation")

text(stigma.sds.fa.varimax$loadings[,1]-0.08, 
     stigma.sds.fa.varimax$loadings[,2]+0.08,
      colnames(stigma.sds),
      col="blue")
abline(h = 0, v = 0)

###########################################################################################################################################

plot(stigma.sds.fa.promax$loadings[,1], 
     stigma.sds.fa.promax$loadings[,2],
     xlab = "Factor 1", 
     ylab = "Factor 2",
     ylim = c(-1,1),
     xlim = c(-1,1),
     main = "Promax rotation")
abline(h = 0, v = 0)

text(stigma.sds.fa.promax$loadings[,1]-0.08, 
     stigma.sds.fa.promax$loadings[,2]+0.08,
      colnames(stigma.sds),
      col="blue")
abline(h = 0, v = 0)

###########################################################################################################################################

### Interpretation:

# Explanation of rotation: 
   #  The purpose of a rotation is to produce factors with a mix of high and low loadings and few moderate-sized loadings. The idea is to give meaning to the 
   #  factors, which helps interpret them. From a mathematical viewpoint, there is no difference between a rotated and unrotated matrix. The fitted model is the
   #  same, the uniqueness are the same, and the proportion of variance explained is the same.


# SDS100 & SDS200 differ from the others, possibly because they are more long term as the others. Contrary to the believe, that the work aspect could also be long term, it might be, that it is not viewed as so long term
# Concluding, it would make sense to combine SDS100 & SDS200 in one index Variable and the rest in one as well.
```

```{r Factor analysis - Social Distance Scale - graph}

## Original fa - Social Distance Scale

str(stigma.sds.fa$loadings)
class(stigma.sds.fa$loadings)

loadings.sds <- xtable(unclass(stigma.sds.fa$loadings)) # putting loadings into data frame
loadings.sds$id <- c("SDS100", "SDS200", "SDS300", "SDS400", "SDS500", "SDS600") # add id

loadings.sds.long <- melt(loadings.sds, id="id", 
                   measure=c("Factor1", "Factor2"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.sds.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)

###########################################################################################################################################

## No Rotation fa - Social Distance Scale

stigma.sds.fa.none <- factanal(stigma.sds, factors = 2, rotation = "none")

loadings.sds.none <- xtable(unclass(stigma.sds.fa.none$loadings)) # putting loadings into data frame
loadings.sds.none$id <- c("SDS100", "SDS200", "SDS300", "SDS400", "SDS500", "SDS600") # add id

loadings.sds.none.long <- melt(loadings.sds.none, id="id", 
                   measure=c("Factor1", "Factor2"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.sds.none.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)

###########################################################################################################################################

## Varimax fa - Social Distance Scale

stigma.ec.fa.varimax <- factanal(stigma.sds, factors = 2, rotation = "varimax")

loadings.ec.varimax <- xtable(unclass(stigma.ec.fa.varimax$loadings)) # putting loadings into data frame
loadings.ec.varimax$id <- c("SDS100", "SDS200", "SDS300", "SDS400", "SDS500", "SDS600") # add id

loadings.ec.varimax.long <- melt(loadings.ec.varimax, id="id", 
                   measure=c("Factor1", "Factor2"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.ec.varimax.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)

###########################################################################################################################################

## Promax fa - Social Distance Scale

stigma.sds.fa.promax <- factanal(stigma.sds, factors = 2, rotation = "promax")

loadings.sds.promax <- xtable(unclass(stigma.sds.fa.promax$loadings)) # putting loadings into data frame
loadings.sds.promax$id <- c("SDS100", "SDS200", "SDS300", "SDS400", "SDS500", "SDS600") # add id

loadings.sds.promax.long <- melt(loadings.sds.promax, id="id", 
                   measure=c("Factor1", "Factor2"), 
                   variable.name="Factor", value.name="Loading")


plot_factor1 <- ggplot(loadings.sds.promax.long, aes(x=factor(id), abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  xlab("Social Distance Scale Questions") +
  scale_x_discrete(labels = c('Neighbour', 'In-Laws', 'Work', 'Food', 'Meeting', 'Child Care')) +
  #scale_y_continuous(labels =c("Neighbour", "Marriage", "Work", "Food", "Meeting", "Child Care")) +
  theme_minimal(base_size=10)

print(plot_factor1)
ggsave(filename = "factorsds.pdf", width = 14, height = 10, dpi = 300, units = "cm")
```


```{r Factor analysis - Attitude Pandemic - FA}

stigma.ec <- df_stigma_long %>%
  select(EC19100, EC19200, EC19300, EC19400)

### Factor Analysis

fa(stigma.ec,nfactors = 2,rotate = "oblimin", fm="minres")
fa.parallel(stigma.ec, fm = "minres", fa = "fa")

# Interpretation: 

# After inspection of the graph it is clear that the variables only load on one factor.

stigma.ec <- na.omit(stigma.ec)
stigma.ec.fa <- factanal(stigma.ec, factors = 1)

### Interpretation the results

apply(stigma.ec.fa$loadings^2,1,sum) # communality (should be high)
1 - apply(stigma.ec.fa$loadings^2,1,sum) # uniquness (should be low)

# factors are worth keeping, as their SS loadings are greater than 1 (Kaiser's rule)


### Residual Matrix

Lambda <- stigma.ec.fa$loadings
Psi <- diag(stigma.ec.fa$uniquenesses)
S <- stigma.ec.fa$correlation
Sigma <- Lambda %*% t(Lambda) + Psi
round(S - Sigma, 6)

```


```{r Factor analysis - Einstellung Lage - Graphs}
### Interpretation Factors

stigma.ec.fa.none <- factanal(stigma.ec, factors = 1, rotation = "none")
stigma.ec.fa.varimax <- factanal(stigma.ec, factors = 1, rotation = "varimax")
stigma.ec.fa.promax <- factanal(stigma.ec, factors = 1, rotation = "promax")


par(mfrow = c(1,3))
plot(stigma.ec.fa.none$loadings[,1], 
     stigma.ec.fa.none$loadings[,1],
     xlab = "Factor 1", 
     ylab = "Factor 1", 
     ylim = c(-1,1),
     xlim = c(-1,1),
     main = "No rotation")
abline(h = 0, v = 0)

###########################################################################################################################################

plot(stigma.ec.fa.varimax$loadings[,1], 
     stigma.ec.fa.varimax$loadings[,1],
     xlab = "Factor 1", 
     ylab = "Factor 1", 
     ylim = c(-1,1),
     xlim = c(-1,1),
     main = "Varimax rotation")

text(stigma.ec.fa.varimax$loadings[,1]-0.08, 
     stigma.ec.fa.varimax$loadings[,1]+0.08,
      colnames(stigma.ec),
      col="blue")
abline(h = 0, v = 0)

###########################################################################################################################################

plot(stigma.ec.fa.promax$loadings[,1], 
     stigma.ec.fa.promax$loadings[,1],
     xlab = "Factor 1", 
     ylab = "Factor 1",
     ylim = c(-1,1),
     xlim = c(-1,1),
     main = "Promax rotation")
abline(h = 0, v = 0)

text(stigma.ec.fa.promax$loadings[,1]-0.08, 
     stigma.ec.fa.promax$loadings[,1]+0.08,
      colnames(stigma.ec),
      col="blue")
abline(h = 0, v = 0)

###########################################################################################################################################

### Interpretation:

# Explanation of rotation: 
   #  The purpose of a rotation is to produce factors with a mix of high and low loadings and few moderate-sized loadings. The idea is to give meaning to the   
   #  factors, which helps interpret them. From a mathematical viewpoint, there is no difference between a rotated and unrotated matrix. The fitted model is the
   #  same, the uniquenesses are the same, and the proportion of variance explained is the same.

# Interpretation:
   #  In the graph we see that EC19100 and EC10400 have similar values. EC19200 and EC19300 have similar values. Which makes sense, because 100 and 400 ask 
   #  after financial situations and the others after health. In general, they have all similar values.
   #  However, the initial factor analysis showed, that all variables only load on one factor, which is why, I am only going to create one Index Variable.

```


```{r Factor analysis - Attitudes towards the pandemic - ggplot2}

## Original fa - Attitude towards the pandemic

str(stigma.ec.fa$loadings)
class(stigma.ec.fa$loadings)

loadings.ec <- xtable(unclass(stigma.ec.fa$loadings)) # putting loadings into data frame
loadings.ec$id <- c("EC19100", "EC19200", "EC19300", "EC19400") # add id

loadings.ec.long <- melt(loadings.ec, id="id", 
                   measure=c("Factor1"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.ec.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)

###########################################################################################################################################

## No Rotation fa - Attitude towards the pandemic

stigma.ec.fa.none <- factanal(stigma.ec, factors = 1, rotation = "none")

loadings.ec.none <- xtable(unclass(stigma.ec.fa.none$loadings)) # putting loadings into data frame
loadings.ec.none$id <- c("EC19100", "EC19200", "EC19300", "EC19400") # add id

loadings.ec.none.long <- melt(loadings.ec.none, id="id", 
                   measure=c("Factor1"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.ec.none.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)

###########################################################################################################################################

## Varimax fa - Attitude towards the pandemic

stigma.ec.fa.varimax <- factanal(stigma.ec, factors = 1, rotation = "varimax")

loadings.ec.varimax <- xtable(unclass(stigma.ec.fa.varimax$loadings)) # putting loadings into data frame
loadings.ec.varimax$id <- c("EC19100", "EC19200", "EC19300", "EC19400") # add id

loadings.ec.varimax.long <- melt(loadings.ec.varimax, id="id", 
                   measure=c("Factor1"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.ec.varimax.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)

###########################################################################################################################################

## Promax fa - Attitude towards the pandemic

stigma.ec.fa.promax <- factanal(stigma.ec, factors = 1, rotation = "promax")

loadings.ec.promax <- xtable(unclass(stigma.ec.fa.promax$loadings)) # putting loadings into data frame
loadings.ec.promax$id <- c("EC19100", "EC19200", "EC19300", "EC19400") # add id

loadings.ec.promax.long <- melt(loadings.ec.promax, id="id", 
                   measure=c("Factor1"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.ec.promax.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)

```


```{r Factor analysis - Attitude towards restrictions}

stigma.mc <- df_stigma_long %>%
  select(MC19100_1, MC19100_2, MC19100_3, MC19100_4, MC19100_5, MC19100_6, MC19100_7, MC19100_8)

### Factor Analysis
stigma.mc <- na.omit(stigma.mc)
stigma.mc.fa <- factanal(stigma.mc, factors = 3)
#stigma.ma.fa <- fa()

### Interpretation the results

apply(stigma.mc.fa$loadings^2,1,sum) # communality (should be high)
1 - apply(stigma.mc.fa$loadings^2,1,sum) # uniquness (should be low)

# factors are worth keeping, as thir SS loadings are greater than 1 (Kaiser's rule)
# did try 4 factors, but the fourth factor was not over 1

### Residual Matrix

Lambda <- stigma.mc.fa$loadings
Psi <- diag(stigma.mc.fa$uniquenesses)
S <- stigma.mc.fa$correlation
Sigma <- Lambda %*% t(Lambda) + Psi
round(S - Sigma, 6)

```


```{r Factor analysis}
### Interpretation Factors

## original fa - Attitude towards restrictions

str(stigma.mc.fa$loadings)
class(stigma.mc.fa$loadings)

loadings.og <- xtable(unclass(stigma.mc.fa$loadings)) # putting loadings into data frame
loadings.og$id <- c("MC19100_1", "MC19100_2", "MC19100_3", "MC19100_4", "MC19100_5", "MC19100_6", "MC19100_7", "MC19100_8") # add id

loadings.og.long <- melt(loadings.og, id="id", 
                   measure=c("Factor1", "Factor2", "Factor3"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.og.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)

###########################################################################################################################################

## no rotation fa - Attitude towards restrictions

stigma.mc.fa.none <- factanal(stigma.mc, factors = 3, rotation = "none")

str(stigma.mc.fa.none$loadings)
class(stigma.mc.fa.none$loadings)

loadings.none <- xtable(unclass(stigma.mc.fa.none$loadings)) # putting loadings into data frame
loadings.none$id <- c("MC19100_1", "MC19100_2", "MC19100_3", "MC19100_4", "MC19100_5", "MC19100_6", "MC19100_7", "MC19100_8") # add id

loadings.none.long <- melt(loadings.none, id="id", 
                   measure=c("Factor1", "Factor2", "Factor3"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.none.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)
###########################################################################################################################################

## varimax fa - Attitude towards restrictions

stigma.mc.fa.varimax <- factanal(stigma.mc, factors = 3, rotation = "varimax")

str(stigma.mc.fa.varimax$loadings)
class(stigma.mc.fa.varimax$loadings)

loadings.varimax <- xtable(unclass(stigma.mc.fa.varimax$loadings)) # putting loadings into data frame
loadings.varimax$id <- c("MC19100_1", "MC19100_2", "MC19100_3", "MC19100_4", "MC19100_5", "MC19100_6", "MC19100_7", "MC19100_8") # add id

loadings.varimax.long <- melt(loadings.varimax, id="id", 
                   measure=c("Factor1", "Factor2", "Factor3"), 
                   variable.name="Factor", value.name="Loading")


ggplot(loadings.varimax.long, aes(id, abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  theme_bw(base_size=10)
###########################################################################################################################################

## promax fa - Attitude towards restrictions

stigma.mc.fa.promax <- factanal(stigma.mc, factors = 3, rotation = "promax")
 
str(stigma.mc.fa.promax$loadings)
class(stigma.mc.fa.promax$loadings)

loadings.promax <- xtable(unclass(stigma.mc.fa.promax$loadings)) # putting loadings into data frame
loadings.promax$id <- c("MC19100_1", "MC19100_2", "MC19100_3", "MC19100_4", "MC19100_5", "MC19100_6", "MC19100_7", "MC19100_8") # add id

loadings.promax.long <- melt(loadings.promax, id="id", 
                   measure=c("Factor1", "Factor2", "Factor3"), 
                   variable.name="Factor", value.name="Loading")


plot_factor2 <- ggplot(loadings.promax.long, aes(x = factor(id), abs(Loading), fill=Loading)) + 
  facet_wrap(~ Factor, nrow=1) + #place the factors in separate facets
  geom_bar(stat="identity") + #make the bars
  coord_flip() + #flip the axes so the test names can be horizontal  
  #define the fill color gradient: blue=positive, red=negative
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint=0, guide=F) +
  ylab("Loading Strength") + #improve y-axis label
  xlab("Different Restrictions to limit the Spread of COVID-19") +
  scale_x_discrete(labels = c('Ban In-Person Schools', 'Ban Private Parties', 'Ban Theatre & Cinema', 'Ban Sports Facilities', 'Limit Social Contact', 'Ban Private Travel', 'Ban Fans at Sport Events', 'Ban Bars & Clubs', 'No Restrictions')) +
  theme_minimal(base_size=10)

print(plot_factor2)
ggsave(filename = "factorres.pdf", width = 14, height = 10, dpi = 300, units = "cm")
###########################################################################################################################################


### Interpretation:

# Explanation of rotation: 
   #  The purpose of a rotation is to produce factors with a mix of high and low loadings and few moderate-sized loadings. The idea is to give meaning to the   
   #  factors, which helps interpret them. From a mathematical viewpoint, there is no difference between a rotated and unrotated matrix. The fitted model is the
   #  same, the uniquenesses are the same, and the proportion of variance explained is the same.

# Interpretation:
   #  According to the factor analysis and the Kaiser's rule, there should be three factors. 
   #  Through the last rotation i would suggest combining 1, 2, 4 and 7. Also 3, 5 and 6. Number 8 loads almost completely on one factor alone, which is why it
   #  should get one exclusive factor. The same for 9, which is not in this factor analysis, as it was exclusiv in the survey and would not make any sense.

```


###########################################################################################################################################
##                                                                                                                                       ##
##                                                 Descriptive Analysis                                                                  ##
##                                                                                                                                       ##
###########################################################################################################################################


```{r , echo=FALSE}

# Vignette Unterschiede:Kontrolle ob die verschiedenen Vignetten unterschiedlich beantwortet wurden über die Vignetten hinweg

ggplot(df_stigma_long, aes(x = gen_sds, y = gen_alter_v)) +
  geom_jitter() + 
  facet_grid(.~ vignr)

ggplot(df_stigma_long, aes(x = gen_sds, y = gen_sex_v)) +
  geom_jitter() + 
  facet_grid(.~ vignr)

ggplot(df_stigma_long, aes(x = gen_sds, y = gen_ethnie_v)) +
  geom_jitter() + 
  facet_grid(.~ vignr)

ggplot(df_stigma_long, aes(x = gen_sds, y = gen_beruf_v)) +
  geom_jitter() + 
  facet_grid(.~ vignr)

ggplot(df_stigma_long, aes(x = gen_sds, y = gen_essential_v)) +
  geom_jitter() + 
  facet_grid(.~ vignr)

ggplot(df_stigma_long, aes(x = gen_sds, y = gen_infektion_v)) +
  geom_jitter() + 
  facet_grid(.~ vignr)


## Interpretation: 

# There seems to be no vignette effect, based on descriptive statistics
```

```{r Boxplots Age}

### Boxplot to look into differences of age vignette

## Age Vignette

ggplot(df_stigma_long, aes(y = gen_sds)) +
  geom_boxplot()  +
  facet_wrap(~gen_alter_v) 

ggplot(df_stigma_long, aes(y = gen_sds_longterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_alter_v)

ggplot(df_stigma_long, aes(y = gen_sds_shortterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_alter_v)

```

```{r Boxplots Gender}

### Boxplot to look into differences of gender vignette

## Gender Vignette

ggplot(df_stigma_long, aes(y = gen_sds)) +
  geom_boxplot()  +
  facet_wrap(~gen_sex_v)

ggplot(df_stigma_long, aes(y = gen_sds_longterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_sex_v)

ggplot(df_stigma_long, aes(y = gen_sds_shortterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_sex_v)
```

```{r Boxplots Ethnicity}

### Boxplot to look into differences of ethnicity vignette

##  Ethnicity Vignette

ggplot(df_stigma_long, aes(y = gen_sds)) +
  geom_boxplot()  +
  facet_wrap(~gen_ethnie_v)

ggplot(df_stigma_long, aes(y = gen_sds_longterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_ethnie_v)

ggplot(df_stigma_long, aes(y = gen_sds_shortterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_ethnie_v)
```

```{r Boxplots occupation}

### Boxplot to look into differences of occupation vignette

## occupation

ggplot(df_stigma_long, aes(y = gen_sds)) +
  geom_boxplot()  +
  facet_wrap(~gen_beruf_v)

ggplot(df_stigma_long, aes(y = gen_sds_longterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_beruf_v)

ggplot(df_stigma_long, aes(y = gen_sds_shortterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_beruf_v)
```

```{r Boxplots Essential}

### Boxplot to look into differences of essential workers

## Essential

ggplot(df_stigma_long, aes(y = gen_sds)) +
  geom_boxplot()  +
  facet_wrap(~gen_essential_v)

ggplot(df_stigma_long, aes(y = gen_sds_longterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_essential_v)

ggplot(df_stigma_long, aes(y = gen_sds_shortterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_essential_v)
```

```{r Boxplots Infection}

### Boxplot to look into differences of infection vignette

## Infection

ggplot(df_stigma_long, aes(y = gen_sds)) +
  geom_boxplot()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(y = gen_sds_longterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(y = gen_sds_shortterm)) +
  geom_boxplot()  +
  facet_wrap(~gen_infektion_v)
```



```{r Hypothesis 2 - Ethnicity}

## Hypothesis 2a and 2b

ggplot(df_stigma_long, aes(x = gen_ethnie_v, y = gen_sds,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_ethnie_v, y = gen_sds_longterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_ethnie_v, y = gen_sds_shortterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)


```

```{r Hypothesis 3 - Occupation}

## Hypothesis 3 - normal difference between all three 

ggplot(df_stigma_long, aes(x = gen_beruf_v, y = gen_sds,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_beruf_v, y = gen_sds_longterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_beruf_v, y = gen_sds_shortterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

## Hypothesis 3 - Difference between essential workfield and non-essential

ggplot(df_stigma_long, aes(x = gen_essential_v, y = gen_sds,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_essential_v, y = gen_sds_longterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_essential_v, y = gen_sds_shortterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

```



```{r Hypothesis 4 - Age}

## Hypothesis 4 - Age

ggplot(df_stigma_long, aes(x = gen_alter_v, y = gen_sds,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_alter_v, y = gen_sds_longterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_alter_v, y = gen_sds_shortterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

```

```{r Control Variable Gender Vignette}

## Gender Vignette

ggplot(df_stigma_long, aes(x = gen_sex_v, y = gen_sds,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_sex_v, y = gen_sds_longterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_sex_v, y = gen_sds_shortterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

```



```{r Control Variable Attitudes towards the pandemic}

## Attitudes towards the pandemic

ggplot(df_stigma_long, aes(x = gen_situation, y = gen_sds,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_situation, y = gen_sds_longterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_situation, y = gen_sds_shortterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

```


```{r Control Variable Attitudes towards restrictions}

## Attitudes towards restrictions

ggplot(df_stigma_long, aes(x = gen_politik, y = gen_sds,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_politik, y = gen_sds_longterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_politik, y = gen_sds_shortterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

```

```{r Control Variable Political Identification}

## Political Identification

ggplot(df_stigma_long, aes(x = gen_herkunft, y = gen_sds,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_herkunft, y = gen_sds_longterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

ggplot(df_stigma_long, aes(x = gen_herkunft, y = gen_sds_shortterm,  alpha = 0.5)) +
  geom_jitter()  +
  facet_wrap(~gen_infektion_v)

```
