==> 04-data-analysis.Rmd <==
---
title: "Analysis"
author: "Sophie Hensgen"
date: "5/28/2021"
output: html_document
---

```{r setup, include=FALSE}
library(readr)
library(tidyverse)
library(dplyr)
library(ggplot2)
library(tidyr)
library(data.table)
library(sem)
library(reshape2)
library(xtable)
library(psych)
library(stargazer)
library(jtools)
library(plm)
library(lmtest)
library(Matrix)
library(lme4)
library(effects)
library(robotstxt)
library(RSocrata)
library(rlist)
library(pipeR)
library(gtrendsR)
library(lubridate)
library(ggpubr)
library(ggiraph)
library(devtools)
library(ggiraphExtra)
library(ggeffects)
library(systemfonts)
library(broom.mixed)
library(sjPlot)
library(sjlabelled)
library(sjmisc)
library(tikzDevice)
library(ggtext)
```

## Load Data Set

```{r Loading Data Set, include = FALSE}

df_stigma_wide <- read_csv("df_stigma_wide.csv")
df_stigma_long <- read_csv("df_stigma_long.csv")

```

## Make Factor Variables

```{r Factor Variables, include = FALSE}

## Variables with more than 2 factors

df_stigma_long$gen_infektion_v <- as.factor(df_stigma_long$gen_infektion_v)
df_stigma_long$gen_ethnie_v <- as.factor(df_stigma_long$gen_ethnie_v)
df_stigma_long$gen_alter_v <- as.factor(df_stigma_long$gen_alter_v)
df_stigma_long$gen_beruf_v <- as.factor(df_stigma_long$gen_beruf_v)
df_stigma_long$gen_politik <- as.factor(df_stigma_long$gen_politik)
df_stigma_long$gen_alter <- as.factor(df_stigma_long$gen_alter)
df_stigma_long$gen_schule_q <- as.factor(df_stigma_long$gen_schule_q)

## Dummy Variables

df_stigma_long$gen_sex_v <- as.factor(df_stigma_long$gen_sex_v)
df_stigma_long$gen_essential_v <- as.factor(df_stigma_long$gen_essential_v)
df_stigma_long$gen_herkunft <- as.factor(df_stigma_long$gen_herkunft) 
df_stigma_long$gen_impfung <- as.factor(df_stigma_long$gen_impfung)
df_stigma_long$gen_bund <- as.factor(df_stigma_long$gen_bund)
df_stigma_long$SD100 <- as.factor(df_stigma_long$SD100)


```

## Changing Reference Category

```{r Reference Category, include = FALSE}

## Variables with more than 2 factors

df_stigma_long$gen_infektion_v <- relevel(df_stigma_long$gen_infektion_v, ref = 3)
df_stigma_long$gen_politik <- relevel(df_stigma_long$gen_politik, ref = 2)
df_stigma_long$gen_schule_q <- relevel(df_stigma_long$gen_schule_q, ref = 1)
df_stigma_long$gen_beruf_v <- relevel(df_stigma_long$gen_beruf_v, ref = 3)
df_stigma_long$SD100 <- relevel(df_stigma_long$SD100, ref = 2)
df_stigma_long$gen_alter <- relevel(df_stigma_long$gen_alter, ref = 1)

```

###########################################################################################################################################
##                                                                                                                                       ##
##                                                 Regression - Random Intercept                                                         ##
##                                                                                                                                       ##
###########################################################################################################################################

## Cluster Robust Standard Error Test

```{r Cluster Robust Standard Error Test, include = FALSE}

## Random Intercept Model - SDS Shortterm

ranin_short_m1 <-  lmer(gen_sds_shortterm ~ 1 + (1 | id), data = df_stigma_long)
summary(ranin_short_m1)

# ICC 0.3578 --> correlated with each other the MCAT scores from any person are with each vignette.
## Explanation --> Imagine if we have students Vignette 1 and 2 from person A; and we have students Vignette 2 and 3 from person B. With a high  P as we have here, it means that Vignette 1 and 2 scores are likely to be similar to each other’s; and Vignette 2 and 3`s scores are likely to be similar to each other’s.

ranin_short_m2 <-  lmer(gen_sds_shortterm ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + (1 | id), data = df_stigma_long)
ranin_short_m3 <-  lmer(gen_sds_shortterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + (1 | id), data = df_stigma_long)
ranin_short_m4 <-  lmer(gen_sds_shortterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter  + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + (1 | id), data = df_stigma_long)
ranin_short_m5 <-  lmer(gen_sds_shortterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter  + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

```

```{r Cluster Robust Standard Error Test, include = FALSE}

## Random Intercept Model - SDS LongTerm

ranin_long_m1 <-  lmer(gen_sds_longterm ~ 1 + (1 | id), data = df_stigma_long)
summary(ranin_long_m1)

# ICC 0.5903 --> correlated with each other the MCAT scores from any person are with each vignette.
## Explanation --> Imagine if we have students Vignette 1 and 2 from person A; and we have students Vignette 2 and 3 from person B. With a high  P as we have here, it means that Vignette 1 and 2 scores are likely to be similar to each other’s; and Vignette 2 and 3`s scores are likely to be similar to each other’s.

ranin_long_m2 <-  lmer(gen_sds_longterm ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + (1 | id), data = df_stigma_long)
ranin_long_m3 <-  lmer(gen_sds_longterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + (1 | id), data = df_stigma_long)
ranin_long_m4 <-  lmer(gen_sds_longterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + (1 | id), data = df_stigma_long)
ranin_long_m5 <-  lmer(gen_sds_longterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

```

```{r Table for LaTeX, include = FALSE}

## Table for LaTeX
 stargazer(ranin_short_m1, ranin_short_m2, ranin_short_m3, ranin_short_m4, ranin_short_m5,ranin_long_m1, ranin_long_m2, ranin_long_m3, ranin_long_m4, ranin_long_m5, 
       title = "Regression Table - Random Intercept Model",
       label="tab2",
       #covariate.labels = c("Infektion", "Female", "Age","Age Squared", "Schulausbildung", "Staatsangehörigkeit", "Politische Einstellung"), 
       #dep.var.labels   = c("Social Distance", "Short Term - Social Distance", "Long Term - Social Distance"), 
       #table.placement = "H",
       column.labels = c("M1", "M2", "M3"),
       model.numbers = FALSE,
       header=FALSE, 
       font.size= "footnotesize", 
       column.sep.width = "-5pt", 
       keep.stat = c("n", "ll", "aic", "bic"))
```

###########################################################################################################################################
##                                                                                                                                       ##
##                                                 Predicted Estimates - Graph                                                           ##
##                                                                                                                                       ##
###########################################################################################################################################

## Predicted Estimate Graph

```{r Hypothesis 1 - predicted Estimates}

## 0. Data set

# head(df_stigma_long) # control data set

## 1. Build linear model 

# 1.1 Build Linear Model DV: SDS Short Term

ranin_short_m5 <-  lmer(gen_sds_shortterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

# 1.2 Build Linear Model DV: SDS Long Term

ranin_long_m5 <-  lmer(gen_sds_longterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

## 2. Plot Model

# 2.1 Use ggpredict to calculate the average predicted estimates

pred_short_h1 <- ggpredict(ranin_short_m5, terms = c("gen_infektion_v"), type = "random", typical = c("typical_value")) # shortterm sds
pred_long_h1 <- ggpredict(ranin_long_m5, terms = c("gen_infektion_v"), type = "random", typical = c("typical_value")) # longterm sds

# 2.2 Plot the data from 2.1 using ggplot2

# create additional variables to make it easier to combine both graphs
pred_long_h1$SDS <- "Long Term SDS"
pred_short_h1$SDS <- "Short Term SDS"
pred_long_h1$SDS <- as.factor(pred_long_h1$SDS)
pred_short_h1$SDS <- as.factor(pred_short_h1$SDS)
pred_h1 <- rbind(pred_long_h1, pred_short_h1)


plot_hyp1 <- pred_h1 %>%
  mutate(x = fct_relevel(x, 
            "1", "2", "3")) %>% # changed order of categorical variable x
  ggplot(aes(x=x, y=predicted, group=SDS, col=SDS)) +
      geom_point() +
      geom_line() + 
      #xlab("Infection Status") +
      #ylab("Predicted Average Estimates") + 
      scale_x_discrete(labels = c('Currently Infected','Former Infected','Not Infected')) +
      scale_color_manual(values=c("#E69F00", "#56B4E9")) +
      #ggtitle( paste("Predicted Infection Status")) +
		  labs( x = "Infection Status", y = "Predicted Average Estimates for SDS") +
      theme_minimal()+
      theme(
      axis.title = element_text(family = 'Times'),
      axis.text.x = element_markdown(family = 'Times'),
      axis.text.y = element_markdown(family = 'Times'),
      legend.text = ggtext::element_markdown(family = 'Times'),
      legend.title = element_markdown(family = 'Times'))
    
print(plot_hyp1)
ggsave(filename = "hyp1.pdf", width = 14, height = 10, dpi = 300, units = "cm")

lm <- lm(gen_sds_shortterm ~ SD100 + gen_bund + gen_mc_largegroups + gen_alter, data =  df_stigma_long)
summary(lm)
```
### Predicted Estimate Graph - Hypothesis 2

```{r Hypothesis 2 - predicted Estimates}

## 0. Data set

# head(df_stigma_long) # control data set

## 1. Build linear model 

# 1.1 Build Linear Model DV: SDS Short Term

ranin_short_m5 <-  lmer(gen_sds_shortterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

# 1.2 Build Linear Model DV: SDS Long Term

ranin_long_m5 <-  lmer(gen_sds_longterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

## 2. Plot Model

# 2.1 Use ggpredict to calculate the average predicted estimates

pred_long_h2 <- ggpredict(ranin_long_m5, terms = c("gen_infektion_v", "gen_ethnie_v"), type = "random", typical = c("typical_value")) # longterm sds

# 2.2 Plot the data from 2.1 using ggplot2

# create additional variables to make it easier to combine both graphs

plot_hyp2 <- pred_long_h2 %>%
  mutate(x = fct_relevel(x, 
            "1", "2", "3")) %>% # changed order of categorical variable x
  ggplot(aes(x=x, y=predicted, group=group, colour = group)) +
      geom_point() +
      geom_line() + 
      xlab("Infection Status") +
      ylab("Predicted Average Estimates") + 
      scale_x_discrete(labels = c('Currently Infected','Former Infected','Not Infected')) +
      scale_color_manual(name = "Ethnicity", values=c("#999999", "#E69F00", "#56B4E9"), labels = c("German", "Turkish", "Asian")) +
      theme_minimal() +
      theme(
      axis.title = element_text(family = 'Times'),
      axis.text.x = element_markdown(family = 'Times'),
      axis.text.y = element_markdown(family = 'Times'),
      legend.text = ggtext::element_markdown(family = 'Times'),
      legend.title = element_markdown(family = 'Times'))
      
print(plot_hyp2)
ggsave(filename = "hyp2.pdf", width = 14, height = 10, dpi = 300, units = "cm")
```


### Predicted Estimate Graph - Hypothesis 3

```{r Hypothesis 3 - predicted Estimates}

## 0. Data set

# head(df_stigma_long) # control data set

## 1. Build linear model 

# 1.1 Build Linear Model DV: SDS Short Term

ranin_short_m5 <-  lmer(gen_sds_shortterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

# 1.2 Build Linear Model DV: SDS Long Term

ranin_long_m5 <-  lmer(gen_sds_longterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

## 2. Plot Model

# 2.1 Use ggpredict to calculate the average predicted estimates

pred_short_h3 <- ggpredict(ranin_short_m5, terms = c("gen_infektion_v", "gen_beruf_v"), type = "random", typical = c("typical_value")) # shortterm sds
pred_long_h3 <- ggpredict(ranin_long_m5, terms = c("gen_infektion_v", "gen_beruf_v"), type = "random", typical = c("typical_value")) # longterm sds

# 2.2 Plot the data from 2.1 using ggplot2

# create additional variables to make it easier to combine both graphs
pred_long_h3$SDS <- "Long Term SDS"
pred_short_h3$SDS <- "Short Term SDS"
pred_long_h3$SDS <- as.factor(pred_long_h3$SDS)
pred_short_h3$SDS <- as.factor(pred_short_h3$SDS)
pred_h3 <- rbind(pred_long_h3, pred_short_h3)

plot_hyp3 <- pred_h3 %>%
  mutate(x = fct_relevel(x, 
            "1", "2", "3")) %>% # changed order of categorical variable x
  ggplot(aes(x=x, y=predicted, group=group, col=group)) +
      geom_point() +
      geom_line() + 
      facet_grid(.~SDS) +
      xlab("Infection Status") +
      ylab("Predicted Average Estimates") + 
      scale_x_discrete(labels = c('Currently Infected','Former Infected','Not Infected')) +
      #scale_color_discrete(name = "Occupation", labels = c("Administrative", "Nurse", "Cashier")) +
      scale_color_manual(name = "Occupation", values=c("#999999", "#E69F00", "#56B4E9"), labels = c("Administrative", "Nurse", "Cashier")) +
      theme_minimal() +
      theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
      theme(
      axis.title = element_text(family = 'Times'),
      axis.text.x = element_markdown(family = 'Times'),
      axis.text.y = element_markdown(family = 'Times'),
      legend.text = ggtext::element_markdown(family = 'Times'),
      legend.title = element_markdown(family = 'Times'))

print(plot_hyp3)
ggsave(filename = "hyp3.pdf", width = 14, height = 10, dpi = 300, units = "cm")
```



### Predicted Estimate Graph - Hypothesis 4

```{r Hypothesis 4 - predicted Estimates}

## 0. Data set

# head(df_stigma_long) # control data set

## 1. Build linear model 

# 1.1 Build Linear Model DV: SDS Short Term

ranin_short_m5 <-  lmer(gen_sds_shortterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

# 1.2 Build Linear Model DV: SDS Long Term

ranin_long_m5 <-  lmer(gen_sds_longterm  ~ gen_infektion_v + gen_alter_v + gen_beruf_v + gen_sex_v + gen_ethnie_v + SD100 + gen_alter + gen_bund + gen_schule_q + gen_herkunft + gen_politik + gen_situation + gen_mc_largegroups + gen_mc_smallgroups + gen_mc_bars + gen_mc_none + gen_impfung + gen_infektion_v*gen_alter_v + gen_infektion_v*gen_beruf_v + gen_infektion_v*gen_ethnie_v + (1 | id), data = df_stigma_long)

## 2. Plot Model

# 2.1 Use ggpredict to calculate the average predicted estimates

pred_long_h2 <- ggpredict(ranin_long_m5, terms = c("gen_infektion_v", "gen_alter_v"), type = "random", typical = c("typical_value")) # longterm sds

# 2.2 Plot the data from 2.1 using ggplot2

# create additional variables to make it easier to combine both graphs

plot_hyp4 <- pred_long_h2 %>%
  mutate(x = fct_relevel(x, 
            "1", "2", "3")) %>% # changed order of categorical variable x
  ggplot(aes(x=x, y=predicted, group=group, colour = group)) +
      geom_point() +
      geom_line() + 
      xlab("Infection Status") +
      ylab("Predicted Average Estimates") + 
      scale_x_discrete(labels = c('Currently Infected','Former Infected','Not Infected')) +
      scale_color_manual(name = "Age Cohort", values=c("#999999", "#E69F00", "#56B4E9"), labels = c("22 Years", "43 Years", "64 Years")) +
      theme_minimal() +
      theme(
      axis.title = element_text(family = 'Times'),
      axis.text.x = element_markdown(family = 'Times'),
      axis.text.y = element_markdown(family = 'Times'),
      legend.text = ggtext::element_markdown(family = 'Times'),
      legend.title = element_markdown(family = 'Times'))
      
print(plot_hyp4)
ggsave(filename = "hyp4.pdf", width = 14, height = 10, dpi = 300, units = "cm")
```
